fix: move config logging tests into existing test file
